Moves BaseBuilder into base_classes.py
